refactor: ファイル名を変更 `carousels` -> `guide-carousel`
