+++ frontend/src/components/features/guide-carousel/guide-carousel.tsx
