+++ frontend/src/components/features/guide-carousel/guide-description.tsx
