+++ frontend/src/components/features/guide-carousel/guide-heading.tsx
