+++ frontend/src/components/features/guide-carousel/guide-image.tsx
