+++ frontend/src/components/features/guide-carousel/guide-step.tsx
